feat: add a chapter selector to tour
